QF-46 replace type TPointsByThemes to current file
