+++ src/types/types.ts
@@ -1,0 +1,9 @@
+type TInfoTheme = {
+    totalPoints: number;
+     totalQuantityQuestions: number
+  };
+  type TPointsByThemes = {
+    [key: string]: TInfoTheme;
+  };
+
+  export type {TPointsByThemes};
